fix: BoardPost 수정 및 styled파일 분리
