+++ src/components/layout/Header.tsx
@@ -1,5 +1,6 @@
 import { useState } from "react";
-import tw from "tailwind-styled-components";
+// import tw from "tailwind-styled-components";
+import tw from "twin.macro";
 import { BsList } from "react-icons/bs";
 import { Link } from "react-router-dom";
 import Logo from "@/components/layout/Logo";
@@ -10,69 +11,71 @@ text-black
 `;
 
 export default function Head() {
-  const [isChecked, setIsChecked] = useState(false);
-  // 이벤트 핸들러 함수 추가
-  const handleDrawerToggle = () => {
-    setIsChecked(!isChecked);
-  };
-  return (
-    <Header>
-      <Link to="/">
-        <Logo></Logo>
-      </Link>
-      <Drawer>
-        <DrawerToggle
-          id="my-drawer"
-          type="checkbox"
-          checked={isChecked}
-          readOnly
-        ></DrawerToggle>
-        <DrawerContent onClick={handleDrawerToggle}>
-          <DrawerButton htmlFor="my-drawer">
-            <ScaledBsList></ScaledBsList>
-          </DrawerButton>
-        </DrawerContent>
-        <DrawerSide>
-          <DrawerOverlay
-            htmlFor="my-drawer"
-            onClick={handleDrawerToggle}
-          ></DrawerOverlay>
-          <SideBarContent>
-            <Link to="/" onClick={handleDrawerToggle}>
-              <Logo></Logo>
+    const [isChecked, setIsChecked] = useState(false);
+    // 이벤트 핸들러 함수 추가
+    const handleDrawerToggle = () => {
+        setIsChecked(!isChecked);
+    };
+    return (
+        <Header>
+            <Link to="/">
+                <Logo></Logo>
             </Link>
-            <Link to="/test" onClick={handleDrawerToggle}>
-              <SiderBarItem className="bg-regal-yellow text-black">
-                테스트 하러가기
-              </SiderBarItem>
-            </Link>
-            <Link to="/stats" onClick={handleDrawerToggle}>
-              <SiderBarItem className="bg-black text-white">
-                통계 보러가기
-              </SiderBarItem>
-            </Link>
-            <Link to="/board" onClick={handleDrawerToggle}>
-              <SiderBarItem className="bg-black text-white">
-                담벼락 보러가기
-              </SiderBarItem>
-            </Link>
-            <SideBarFooter>
-              <SideBarFooterContent className="text-black">AYT Company</SideBarFooterContent>
-              <SideBarFooterContent className="pointer-events-auto text-black">
-                <a
-                  href="https://github.com/are-you-T"
-                  target="_blank"
-                  rel="are-you-T noreferrer"
-                >
-                  https://github.com/are-you-T
-                </a>
-              </SideBarFooterContent>
-            </SideBarFooter>
-          </SideBarContent>
-        </DrawerSide>
-      </Drawer>
-    </Header>
-  );
+            <Drawer>
+                <DrawerToggle
+                    id="my-drawer"
+                    type="checkbox"
+                    checked={isChecked}
+                    readOnly
+                ></DrawerToggle>
+                <DrawerContent onClick={handleDrawerToggle}>
+                    <DrawerButton htmlFor="my-drawer">
+                        <ScaledBsList></ScaledBsList>
+                    </DrawerButton>
+                </DrawerContent>
+                <DrawerSide>
+                    <DrawerOverlay
+                        htmlFor="my-drawer"
+                        onClick={handleDrawerToggle}
+                    ></DrawerOverlay>
+                    <SideBarContent>
+                        <Link to="/" onClick={handleDrawerToggle}>
+                            <Logo></Logo>
+                        </Link>
+                        <Link to="/test" onClick={handleDrawerToggle}>
+                            <SiderBarItem className="bg-regal-yellow text-black">
+                                테스트 하러가기
+                            </SiderBarItem>
+                        </Link>
+                        <Link to="/stats" onClick={handleDrawerToggle}>
+                            <SiderBarItem className="bg-black text-white">
+                                통계 보러가기
+                            </SiderBarItem>
+                        </Link>
+                        <Link to="/board" onClick={handleDrawerToggle}>
+                            <SiderBarItem className="bg-black text-white">
+                                담벼락 보러가기
+                            </SiderBarItem>
+                        </Link>
+                        <SideBarFooter>
+                            <SideBarFooterContent className="text-black">
+                                AYT Company
+                            </SideBarFooterContent>
+                            <SideBarFooterContent className="pointer-events-auto text-black">
+                                <a
+                                    href="https://github.com/are-you-T"
+                                    target="_blank"
+                                    rel="are-you-T noreferrer"
+                                >
+                                    https://github.com/are-you-T
+                                </a>
+                            </SideBarFooterContent>
+                        </SideBarFooter>
+                    </SideBarContent>
+                </DrawerSide>
+            </Drawer>
+        </Header>
+    );
 }
 
 const Header = tw.header`
@@ -154,7 +157,6 @@ const SideBarFooter = tw.li`
 `;
 
 const SideBarFooterContent = tw.div`
-  flex-end
   self-center
 `;
 // border-radius: 100px;
